--- starter-code/routes/secreto.js
@@ -1,9 +1,0 @@
-const express = require('express');
-const isLoggedIn = require('../middlewares/isLoggedIn');
-const router = express.Router();
-
-router.get('/',isLoggedIn, (req,res) => {
-  res.render('secreto');
-});
-
-module.exports = router;
