--- starter-code/views/secreto.ejs
@@ -1,1 +1,0 @@
-<p>Hola</p>
